requirements.txt
fastapi
uvicorn
pandas
spacy
transformers
torch
pinecone-client
numpy
openpyxl